% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_cluster_validity.R
\name{avgSilhouette}
\alias{avgSilhouette}
\title{avgSilhouette}
\usage{
avgSilhouette(clusters, dmatrix)
}
\arguments{
\item{clusters}{Vector where values are cluster assignments for samples}

\item{dmatrix}{Dissimilarity matrix as generated by the
\code{dist()} function}
}
\value{
Numeric value
}
\description{
For a clustering solution for any k, compute the average silhouette
width as an internal measure of cluster validity
}
