% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromswitch.R
\docType{package}
\name{chromswitch}
\alias{chromswitch}
\alias{chromswitch-package}
\title{chromswitch: An R package for detecting chromatin state switches}
\description{
chromswitch implements a flexible method to detect chromatin state
switches between samples in two biological conditions in a specific genomic
region of interest given peaks called from ChIP-seq data.
}
