% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster}
\alias{cluster}
\title{cluster}
\usage{
cluster(ft_mat, metadata, region, heatmap = TRUE, title = NULL)
}
\arguments{
\item{ft_mat}{matrix where columns are features and rows are samples as
returned by \code{\link{summarizePeaks}} or \code{\link{positionAware}}}

\item{metadata}{A dataframe with a column "Sample" which stores
the sample identifiers, and at least one column, "Group", which stores
the biological condition labels of the samples}

\item{region}{GRanges object specifying the query region}

\item{heatmap}{(Optional) Logical value indicating whether or not to plot
the heatmap for hierarchical clustering}

\item{title}{(Optional) If \code{heatmap} is TRUE, specify the title of the
plot, which will also be used for the output file name in PDF format}
}
\value{
List with two elements: "stats" is a one-row dataframe with cluster
validity statistics for the clustering result, "clusters" is a character
vector where names are sample IDs and values are cluster labels
}
\description{
cluster
}
\examples{
samples <- c("brain1", "brain2", "brain3", "other1", "other2", "other3")
outfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")
groups <- c(rep("Brain", 3), rep("Other", 3))

metadata <- data.frame(Sample = samples,
    H3K4me3 = outfiles,
    Group = groups,
    stringsAsFactors = FALSE)

region <- GenomicRanges::GRanges(seqnames = "chr19",
    ranges = IRanges::IRanges(start = 54924104, end = 54929104))

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = region)

ft_mat <- summarizePeaks(lpk, mark = "H3K4me3",
cols = c("qValue", "signalValue"))

cluster(ft_mat, metadata, region)

}
