% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_cluster_validity.R
\name{clusterValidityPerK}
\alias{clusterValidityPerK}
\title{clusterValidityPerK}
\usage{
clusterValidityPerK(ft_mat)
}
\arguments{
\item{ft_mat}{Data frame or matrix, feature matrix as generated by
\code{\link{summarizePeaks}} or \code{\link{binarizePeaks}}}
}
\value{
Data frame with one row per value of k
}
\description{
Given an hclust object and the corresponding dissimilarity
matrix, cut the tree at every possible height \code{k} and compute internal
measures of cluster validity. For now, we compute only the Average Silhouette
width of the partition.
}
