% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_cluster_validity.R
\name{completeness}
\alias{completeness}
\title{completeness}
\usage{
completeness(contingency, c, k)
}
\arguments{
\item{contingency}{Table, contingency table between clusters and
conditions as returned by the \code{table} function}

\item{c}{Vector of classes}

\item{k}{Vector of clusters}
}
\value{
Numeric
}
\description{
Computes the completeness of a set of clusters given ground-truth
classes, as defined in
https://aclweb.org/anthology/D/D07/D07-1043.pdf
}
\examples{

clusters <- c(0, 0, 2, 1, 1, 0, 1)
classes <- c("A", "A", "A", "B", "B", "A", "B")
ct <- table(classes, clusters)
completeness(contingency = ct)

}
