% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_cluster_validity.R
\name{conditionalClassEntropy}
\alias{conditionalClassEntropy}
\title{classEntropyGivenClusters}
\usage{
conditionalClassEntropy(contingency, c, k)
}
\arguments{
\item{contingency}{Table, contingency table between clusters and
conditions as returned by the \code{table} function}

\item{c}{Vector of classes}

\item{k}{Vector of clusters}
}
\value{
Numeric
}
\description{
Computes the conditional entropy of a set of classes, given the
cluster assignments, as defined in
https://aclweb.org/anthology/D/D07/D07-1043.pdf
}
