% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_cluster_validity.R
\name{contingencyTable}
\alias{contingencyTable}
\title{contingencyTable}
\usage{
contingencyTable(clusters, metadata)
}
\arguments{
\item{clusters}{A vector where values are the cluster assignments for each
sample}

\item{metadata}{Data frame with at least two columns: "Sample" storing
sample IDs and "Condition" storing biological conditions of each samples}
}
\value{
table
}
\description{
For each assigned cluster, stores the number of samples which belong
to each known class. Used for evaluating the clustering solution. This
function exists to create the contingencyTable from the metadata, to
ensure that the clusters & conditions for the classes are associated
correctly based on sample IDs.
}
\keyword{internal}
