% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordToGRanges}
\alias{coordToGRanges}
\title{coordToGRanges}
\usage{
coordToGRanges(coord)
}
\arguments{
\item{coord}{String coordinate in genome browser-friendly format to convert
to a GRanges object}
}
\value{
GRanges object
}
\description{
Convert a string of genomic coordinates to a GRanges object
}
\examples{
string <- "chr1:1000-2000"
coordToGRanges(string)

}
