% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_processing.R
\name{filterPeaks}
\alias{filterPeaks}
\title{filterPeaks}
\usage{
filterPeaks(peaks, columns, thresholds)
}
\arguments{
\item{peaks}{List of GRanges objects, as returned by \code{\link{loadBed}}}

\item{columns}{Character vector of column names containing stats by which to
filter peaks}

\item{thresholds}{Vector of numeric values giving the lower thresholds to use
for each of the columns specified, in the same order as \code{columns}}
}
\value{
A list of GRanges objects storing peak calls for each sample, with
peaks filtered according to the columns and thresholds specified.
}
\description{
Given a set of peak calls for different marks and samples as returned by
\code{\link{loadBed}}, filter peaks according to values in numeric
}
\examples{
filterPeaks(peaks = H3K4me3,
    columns = c("signalValue", "pValue"),
    thresholds = c(4, 10))

}
