% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_processing.R
\name{filterSamplePeaks}
\alias{filterSamplePeaks}
\title{filterSamplePeaks}
\usage{
filterSamplePeaks(sample_peaks, columns, thresholds)
}
\arguments{
\item{sample_peaks}{GRanges object (peak calls for one sample)}

\item{columns}{Character vector of column names containing stats by which to
filter peaks}

\item{thresholds}{Vector of numeric values giving the lower thresholds to use
for each of the columns specified, in the same order as \code{columns}}
}
\value{
A GRanges object containing peak calls for this sample, filtered
according to the thresholds and columns specified.
}
\description{
filterSamplePeaks
}
\seealso{
filterPeaks
}
\keyword{internal}
