% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_aware_matrix.R
\name{getSamplePeakProfile}
\alias{getSamplePeakProfile}
\title{getSamplePeakProfile}
\usage{
getSamplePeakProfile(peaks, windows, p)
}
\arguments{
\item{peaks}{GRanges object containing peaks for one sample}

\item{windows}{GRanges object containing windows in which to model presence/
absence of peaks, these become the features or columns of the output
or columns}

\item{p}{Numeric value in [0, 1] giving the fraction of reciprocal overlap
to require.}
}
\value{
A one-row dataframe with a logical value TRUE/FALSE in each column
(window) indicating whether any peaks overlap the window
}
\description{
For a set of peaks in one sample and a set of windows, which could correspond
to bins of a region, or unique peaks, model presence or absence of peaks in
each window by calling peaks present in a window if peaks in the sample have
a p-reciprocal overlap with the window. Alternatively, the windows can be
viewed as unique peaks, and this function calls the presence or absence of
each peak in the sample.
}
\keyword{internal}
