% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_aware_matrix.R
\name{getUniquePeaks}
\alias{getUniquePeaks}
\title{getUniquePeaks}
\usage{
getUniquePeaks(loc_union, p)
}
\arguments{
\item{loc_union}{GRanges object storing union of peaks in the query
region across samples}

\item{p}{Numeric value in [0, 1] giving the fraction of reciprocal overlap
to require.}
}
\value{
GRanges objects with a list of unique peaks as determined by
the p-reciprocal overlap rule
}
\description{
Given a set of peaks, collapse them such that we retain only unique peaks
which do not have a reciprocal overlap of \code{p}%. We do so by scanning
through the union of peaks across samples in the query region in the order
they're provided, and growing a list of unique peaks. For every peak visited,
add it to the list of unique peaks if it doesn't have a p-reciprocal overlap
with any peak already in the list;E otherwise, discard it.
}
