% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_cluster_validity.R
\name{internalClusterValidity}
\alias{internalClusterValidity}
\title{clusterValidity}
\usage{
internalClusterValidity(k, hclust_obj, dmatrix)
}
\arguments{
\item{k}{Numeric value giving number of clusters, used to cut the dendrogram
from an \code{hclust} object}

\item{hclust_obj}{hclust object}

\item{dmatrix}{Dissimilarity matrix as produced by the \code{dist} function}
}
\value{
Data frame with one row
}
\description{
For a given value of k, an hclust object and the corresponding dissimilarity
matrix, cut the tree at height \code{k} and compute internal measures of
cluster validity. For now, we compute only the Average Silhouette width
of the partition.
}
