% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPeaks-class.R
\docType{methods}
\name{is.empty}
\alias{is.empty}
\alias{is.empty,localPeaks-method}
\title{is.empty}
\usage{
is.empty(object)

\S4method{is.empty}{localPeaks}(object)
}
\arguments{
\item{object}{The object to check for emptiness}
}
\value{
Logical value indicating if the object is empty (TRUE) or not (FALSE)
}
\description{
is.empty
}
\section{Methods (by class)}{
\itemize{
\item \code{localPeaks}: Returns TRUE if the localPeaks object has no peaks in
any of the samples in \code{object@peaks}, i.e. if no peaks were found in
the query region.
}}

\keyword{internal}
