% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{loadSampleBed}
\alias{loadSampleBed}
\title{loadSampleBed}
\usage{
loadSampleBed(bed_file, metadata_cols = NULL)
}
\arguments{
\item{bed_file}{Character string containing path to BED file with first three
columns "chrm", "start", "end" and optional additional columns}

\item{metadata_cols}{Optional: a character vector containing names of
all additional columns. If not specified but additional columns are found,
they are dropped. If some are specified, it is assumed that they describe
the columns which immediately follow "end".}
}
\value{
A GRanges object
}
\description{
Given a path, read a BED file into a dataframe
}
