% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPeaks-class.R
\name{localPeaks}
\alias{localPeaks}
\title{localPeaks object}
\usage{
localPeaks(region, peaks, samples)
}
\arguments{
\item{region}{A GRanges object specifying one genomic region,
the query region}

\item{peaks}{List of lists of GRanges objects. Each outer list stores peaks
for each sample for one mark in \code{region}.}

\item{samples}{Character vector with sample identifiers.}
}
\value{
localPeaks object
}
\description{
A localPeaks object is a container for the peaks for one or more marks for
a set of samples in a specific genomic region of interest, as well as the
genomic region itself, and the sample IDs. These components are needed to
convert sets of peaks into rectangular feature-by-sample matrices which we
can then use for downstream analysis - and in particular, as input to a
clustering algorithm in order to call a chromatin state switch. This function
is a constructor for a localPeaks object.
}
\keyword{internal}
