% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPeaks-class.R
\name{lpkPeaks}
\alias{lpkPeaks}
\title{lpkPeaks}
\usage{
lpkPeaks(lpks)
}
\arguments{
\item{lpks}{localPeaks object}
}
\value{
List of lists of GRanges objects. Each outer list stores peaks
for each sample for one mark in the region given by \code{lpkRegion(lpks)}.
}
\description{
Accessor for \code{peaks} slot of a \code{\link{localPeaks}} object.
}
\examples{
samples <- c("E068", "E071", "E074", "E101", "E102", "E110")
outfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = outfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GRanges(seqnames = "chr19",
    ranges = IRanges(start = 54924104, end = 54929104)))

lpkPeaks(lpk)

}
