% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPeaks-class.R
\name{lpkRegion}
\alias{lpkRegion}
\title{lpkRegion}
\usage{
lpkRegion(lpks)
}
\arguments{
\item{lpks}{localPeaks object}
}
\value{
GRanges object with query region associated with the localPeaks
object
}
\description{
Accessor for \code{region} slot of a \code{\link{localPeaks}} object.
}
\examples{

samples <- c("brain1", "brain2", "brain3", "other1", "other2", "other3")
outfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = outfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GenomicRanges::GRanges(seqnames = "chr19",
    ranges = IRanges::IRanges(start = 54924104, end = 54929104)))

lpkRegion(lpk)

}
