% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{makeBrowserCoord}
\alias{makeBrowserCoord}
\title{makeBrowserCoord}
\usage{
makeBrowserCoord(chr, start, end)
}
\arguments{
\item{chr}{The chromosome}

\item{start}{The starting position of the genomic region}

\item{end}{The ending position of the genomic region}
}
\value{
String with copy-pastable, genome browser-friendly version of
coordinates.
}
\description{
Given coordinates for a genomic region, return a browser-friendly version.
}
\examples{

makeBrowserCoord("chr1", 1000, 2000)

}
