% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_cluster_validity.R
\name{NMI}
\alias{NMI}
\title{NMI}
\usage{
NMI(clusters, classes)
}
\arguments{
\item{clusters}{A vector of cluster assignments}

\item{classes}{A vector giving the true classes of the objects}
}
\value{
Numeric
}
\description{
Computes the Normalized Mutual Information betwen two partitions
}
\details{
This code comes directly from the package `clue`:
\url{https://github.com/cran/clue/blob/098da43010f3803294b4e
8403328ee5c0216abf7/R/agreement.R#L161}

Hornik K (2017). _clue: Cluster ensembles_. R package version
0.3-53, <URL: https://CRAN.R-project.org/package=clue>.

Hornik K (2005). “A CLUE for CLUster Ensembles.” _Journal of
Statistical Software_, *14*(12). doi: 10.18637/jss.v014.i12 (URL:
http://doi.org/10.18637/jss.v014.i12).
}
