% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_processing.R
\name{normalizePeaks}
\alias{normalizePeaks}
\title{normalizePeaks}
\usage{
normalizePeaks(peaks, columns, tail = 0.005)
}
\arguments{
\item{peaks}{List of GRanges objects storing peak calls for each sample,
where element names correspond to sample IDs}

\item{columns}{Character vector specifying the names of columns to normalize}

\item{tail}{Optional: numeric, a fraction in [0, 1] specifying how much of
the data to bound to 0 (for the lower tail) or 1 (for the upper tail).
Default: 0.005.}
}
\value{
A list of GRanges objects storing peak calls for each sample, with
columns specified in \code{columns} normalized.
}
\description{
Given a set of peak calls for different marks and samples, normalize all
peaks genome-wide for each
sample and mark by rescaling and Winsorizing,
i.e. rescale the middle of the data to the range
[0, 1] and bound the upper tail to 1 and the lower tail to 0, effectively
replacing a fixed amount of extreme values in each tail. Similar to trimming
the tails except instead of discarding the tails entirely they're bounded.
}
\examples{
normalizePeaks(H3K4me3, columns = c("signalValue", "pValue", "qValue"))

}
\seealso{
winsorNorm
}
