% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_region_matrix.R
\name{peakOverlap}
\alias{peakOverlap}
\title{peakOverlap}
\usage{
peakOverlap(region, peaks)
}
\arguments{
\item{region}{GRanges object defining query region}

\item{peaks}{GRanges object storing peak calls for a sample in the
query region}
}
\value{
double, a number in [0,1].
}
\description{
Compute the fraction of the query region of interest which is overlapped
by peaks. This value will be at most 1, since we only consider the
overlapping portion of each peak, and peaks do not overlap with each other.
This value will be 0 in the case that there are no peaks in the query region.
}
