% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalPeaks-class.R
\docType{methods}
\name{peaks,LocalPeaks-method}
\alias{peaks,LocalPeaks-method}
\alias{peaks-method}
\title{peaks}
\usage{
\S4method{peaks}{LocalPeaks}(x)
}
\arguments{
\item{x}{LocalPeaks object}
}
\value{
List of lists of GRanges objects. Each outer list stores peaks
for each sample for one mark in the region given by \code{region(lpks)}.
}
\description{
Accessor for \code{peaks} slot of a \code{\linkS4class{LocalPeaks}} object.
}
\examples{
samples <- c("E068", "E071", "E074", "E101", "E102", "E110")
bedfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = bedfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GRanges(seqnames = "chr19",
    ranges = IRanges(start = 54924104, end = 54929104)))

peaks(lpk)

}
