% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_aware_matrix.R
\name{positionAware}
\alias{positionAware}
\title{positionAware}
\usage{
positionAware(lpks, p)
}
\arguments{
\item{lpks}{localPeaks object storing peaks for all samples in the query
region}

\item{p}{Numeric value in [0, 1] giving the fraction of reciprocal overlap
to require.}
}
\value{
A data frame where rows are samples and columns are features. The
genomic ranges which give the features are returned as the \code{features}
attribute of the data frame.
}
\description{
Given peaks for a set of samples in a query region, construct a sample-by-
feature matrix where each row is a binary vector which models the presence
or absence of unqiue peaks in the region.
}
\examples{
samples <- c("brain1", "brain2", "brain3", "other1", "other2", "other3")
outfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = outfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GenomicRanges::GRanges(seqnames = "chr19",
    ranges = IRanges::IRanges(start = 54924104, end = 54929104)))

# Get feature matrix
ft_matrix <- positionAware(lpk, 0.5)

# See features
attr(ft_matrix, "features")

}
