% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_aware_matrix.R
\name{pReciprocalOverlap}
\alias{pReciprocalOverlap}
\title{pReciprocalOverlap}
\usage{
pReciprocalOverlap(a, b, p)
}
\arguments{
\item{a}{GRanges object storing first region}

\item{b}{GRanges object storing second region}

\item{p}{Numeric value in [0, 1] giving the fraction of reciprocal overlap
to require.}
}
\value{
Logical value, TRUE if \code{a} and \code{b} are the same
by having a p-reciprocal overlap, FALSE otherwise
}
\description{
If a and b denote two genomic regions, check whether they overlap
reciprocally by \code{p}*100% of their lengths
}
\examples{
a <- GenomicRanges::GRanges(seqnames = "chr11",
            ranges = IRanges::IRanges(start = 112829468, end = 112834468))

b <- GenomicRanges::GRanges(seqnames = "chr11",
            ranges = IRanges::IRanges(start = 112829468, end = 113834468))

c <- GenomicRanges::GRanges(seqnames = "chr11",
            ranges = IRanges::IRanges(start = 112829968, end = 112834968))

# Expect FALSE
pReciprocalOverlap(a, b, 0.9)

# Expect TRUE
pReciprocalOverlap(a, c, 0.9)

}
