% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_cluster_validity.R
\name{purity}
\alias{purity}
\title{purity}
\usage{
purity(contingency, c, k)
}
\arguments{
\item{contingency}{Table, contingency table between clusters and
conditions as returned by the \code{table} function}

\item{c}{Vector of classes}

\item{k}{Vector of clusters}
}
\value{
Numeric
}
\description{
Computes the purity of a partition as defined in
https://www.ncbi.nlm.nih.gov/pubmed/17483501
}
