% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalPeaks-class.R
\docType{methods}
\name{region,LocalPeaks-method}
\alias{region,LocalPeaks-method}
\alias{region-method}
\title{region}
\usage{
\S4method{region}{LocalPeaks}(x)
}
\arguments{
\item{x}{LocalPeaks object}
}
\value{
GRanges object with query region associated with the LocalPeaks
object
}
\description{
Accessor for \code{region} slot of a \code{\linkS4class{LocalPeaks}} object.
}
\examples{

samples <- c("E068", "E071", "E074", "E101", "E102", "E110")
bedfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = bedfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GRanges(seqnames = "chr19",
    ranges = IRanges(start = 54924104, end = 54929104)))

region(lpk)

}
