% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_processing.R
\name{retrieveSamplePeaks}
\alias{retrieveSamplePeaks}
\title{retrieveSamplePeaks}
\usage{
retrieveSamplePeaks(peaks, region)
}
\arguments{
\item{peaks}{GRanges object containing peak calls for the sample}

\item{region}{GRanges object specifying genomic region in which to search
for peaks}
}
\value{
GRanges object. Contains all peaks overlapping query region.
}
\description{
Given a set of peaks for one sample and a region of interest,
retrieve peaks overlapping query region
}
\keyword{internal}
