% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_matrix.R
\name{summarizeSamplePeaks}
\alias{summarizeSamplePeaks}
\title{summarizeSamplePeaks}
\usage{
summarizeSamplePeaks(peaks, region, mark, cols, length = FALSE,
  fraction = TRUE, n = FALSE)
}
\arguments{
\item{peaks}{GRanges object storing peak calls for a sample in the
query region}

\item{region}{GRanges object storing query region}

\item{mark}{String specifying the name of the mark for which peaks are given}

\item{cols}{Character vector of column names on which to compute summary
statistics}

\item{length}{Logical: compute the mean, median, and max of peak length?}

\item{fraction}{Loogical: compute the fraction of the region overlapped by
peaks?}

\item{n}{Logical: compute the number of peaks in the region?}
}
\value{
data.frame
}
\description{
For one sample, given peaks in the query region, compute several summary
statistics of peaks in that region, including the mean, median, and max of
any specified
}
\keyword{internal}
