% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_processing.R
\name{winsorNorm}
\alias{winsorNorm}
\title{winsorNorm}
\usage{
winsorNorm(x, trim)
}
\arguments{
\item{x}{A numeric vector, the data to be normalized}

\item{trim}{Numeric, a fraction in [0, 1] specifying how much of the data
to bound to 0 (for the lower tail) or 1 (for the upper tail)}
}
\value{
Numeric vector
}
\description{
Normalize a numeric vector by rescaling and Winsorizing, i.e. rescale the
middle of the data to the range [0, 1] and bound the upper tail to 1 and the
lower tail to 0, effectively replacing a fixed amount of extreme values in
each tail. Similar to trimming the tails except instead of discarding
the tails entirely they're bounded.
}
\examples{

x <- seq(1, 100, by = 1)
x

# Bound the lower and upper 5\% of values in the vector
winsorNorm(x, trim = 0.05)

}
