% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPeaks-class.R
\docType{methods}
\name{samples,localPeaks-method}
\alias{samples,localPeaks-method}
\alias{samples-method}
\title{samples}
\usage{
\S4method{samples}{localPeaks}(object)
}
\arguments{
\item{object}{localPeaks object}
}
\value{
Character vector with sample IDs for the localPeaks object
}
\description{
Accessor for \code{samples} slot of a \code{\linkS4class{localPeaks}} object.
}
\examples{
samples <- c("E068", "E071", "E074", "E101", "E102", "E110")
bedfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = bedfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GRanges(seqnames = "chr19",
    ranges = IRanges(start = 54924104, end = 54929104)))

samples(lpk)

}
