% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_region_matrix.R
\name{summarizePeaks}
\alias{summarizePeaks}
\title{summarizePeaks}
\usage{
summarizePeaks(localpeaks, mark, cols, length = FALSE, fraction = TRUE,
  n = FALSE)
}
\arguments{
\item{localpeaks}{localPeaks object}

\item{mark}{String specifying the name of the mark for which the localPeaks
object is given}

\item{cols}{Character vector of column names on which to compute summary
statistics}

\item{length}{Logical: compute the mean, median, and max of peak length?}

\item{fraction}{Loogical: compute the fraction of the region overlapped by
peaks?}

\item{n}{Logical: compute the number of peaks in the region?}
}
\value{
A matrix where rows are samples and columns are features
}
\description{
Given peaks for a set of samples in a query region, construct a sample-by-
feature matrix where each row is a vector of summary statistics computed from
peaks in the region.
}
\examples{
samples <- c("brain1", "brain2", "brain3", "other1", "other2", "other3")
outfiles <- system.file("extdata", paste0(samples, ".H3K4me3.bed"),
package = "chromswitch")

metadata <- data.frame(Sample = samples,
    H3K4me3 = outfiles,
    stringsAsFactors = FALSE)

lpk <- retrievePeaks(H3K4me3,
    metadata = metadata,
    region = GenomicRanges::GRanges(seqnames = "chr19",
    ranges = IRanges::IRanges(start = 54924104, end = 54929104)))

summarizePeaks(lpk, mark = "H3K4me3", cols = c("qValue", "signalValue"))

}
